fixed bug in version handling code

+++ src/version.c
@@ -3,7 +3,7 @@
 
 // prints the buildnumber to stdout
 int main() {
-	unsigned long long buildNumber;
+	unsigned long buildNumber;
 	FILE *f;
 	if ((f = fopen("./var/version.cfg","r")) == NULL) { // if no version file exists, create one
 		f = fopen("./var/version.cfg","w");
@@ -11,7 +11,7 @@ int main() {
 		fprintf(f,"%s %d\n",VERSION_STRING,buildNumber);
 	} else { // a version file exists, so use it
 		char *version = MALLOC_STRING;
-		fscanf(f, "%s %llu",version,&buildNumber);
+		fscanf(f, "%s %lu",version,&buildNumber);
 		fclose(f);
 		f = fopen("./var/version.cfg","w");
 		if (strcmp(version,VERSION_STRING) == 0) { // if we're still on the same version
@@ -21,6 +21,6 @@ int main() {
 		}
 		fprintf(f,"%s %d\n",VERSION_STRING,buildNumber);
 	}
-	printf("%llu",buildNumber);
+	printf("%lu",buildNumber);
 	return 0;
 }
